Fixed GTK+ rendering issues and limited xz's memory consumption.
GTK+ cannot render text or images, because the Pango and GDK-PixBuf modules
cache is missing. This bug was fixed by expanding nscd's init script, so it also
generates it.

In addition, xz's memory usage limit is now set through /etc/profile.d/xz.sh
instead of the default, hardcoded value. In the case of Slackware 14.0, buildpkg
fails a lot because xz reaches the low memory limit. This makes that limit
configurable.

Besides these issues, Subito GNU/Linux-libre seems fast and solid, as it should
be.