Built-in support for sources download in buildpkg and a kernel sources package.
download() is still implemented by build scripts, but only for anything other
than downloading a file to the working directory, under its original name: for
instance, all Git or SVN snapshots are still downloaded via download(). In the
vast majority of the build scripts shipped with roar-ng II, the script just
downloads one file, so this is a nice infrastructure feature.

In addition, a kernel sources package was added. From now and on, both the
kernel itself and the kernel headers package are built using the kernel sources
in /usr/src/linux, which are the contents of the new kernel sources package. In
all kernel packages, the package revision is now changed to the minor version of
the kernel, while the package version is the major version, as done in some
major distributions.

This is nice to have for three reasons:
  - The build scripts can share the same sources, so extraction, patching and
    deblobbing are done only once. The kernel can be rebuilt quickly, without
    having to go through the annoying sources modification process again and
    again.
  - Third-party drivers can be built more easily.
  - It's extremely convenient.