Added support for zram, on by default.
zram is a compressed block device located in RAM that is used for swap. Tiny
Core and Slax already utilize it and users seem to like it.

The roar-ng II implementation follows the author's scalability notes - the
number of devices created is configurable, and set to the number of available
processors, by default, while the default size for a single device is 128 MB.

This support is provided by an init script, for two reasons:
  - If a machine lacks the amount of memory required to boot a distribution,
    zram doesn't solve the problem. It makes it boot, but doesn't add RAM -
    that's the real issue.
  - It should be an optional feature that can be disabled easily.