Added Puppy Linux support.
Due to the dirty nature of the Puppy Package Manager (PPM), this support is
partial. Package lists of Puppy Linux do not have an architecture field;
instead, each repository contains packages of one architecture, without any
means to guess what it is in a generic and clean way.

In order to use Puppy packages for multiple architectures, one must add their
repositories to the list in this plug-in.