Merged subito-desktop into the skeleton.
From now and on, each distribution built using roar-ng II will use the same
desktop infrastructure as Subito GNU/Linux-libre, unless the developer changes
skeleton.

Files added to the skeleton are:
  - Configuration files for Vim, GNU Screen, cwm and Bash.
  - X Server configuration files: .Xdefaults and .xinitrc.

The latter depends on rxvt-unicode, xsetroot and cwm: it sets the root window
color, starts urxvtd and launches cwm. It should be replaced with another, in
order to provide a different startup process for the graphical environment,
through the window manager package or a package similar to the deprecated
"subito-desktop".