Replaced re-alpine with Mutt, msmtp and Abook.
This combination is much lighter, but the main reason it's used instead of
re-alpine is the fact the latter fails to build on Debian and derivatives,
unless modified.

In addition, since both Mutt and re-alpine are not maintained well, Mutt is more
secure when used with external (e.g msmtp), well-maintained and secure tools.