Bug fix - the Perl modules directory is a symlink under Debian and derivatives.
This symlink causes problems with third-party Perl modules, such as Irssi's.
Since developers assume the modules directory is a directory (and not a
symlink), external modules are installed in a directory and get copied into a
symlink when roar-ng II builds the root file system.

This commit solves this problem by replacing the symlink with a directory. This
is safe, because only one minor version of each major Perl version can be
installed at the same time.