Added support for alternate package names.
From now and on, each roar-ng II package may have more than one name. This is
needed, since some major distributions (most notably, Debian and CentOS) use a
"provides" field for each package, which tells which components (e.g the
system's default shell) it provides. These components are used as alternate
names for the package, in the dependency lists of other packages.

In addition, three changes are introduced in the Debian support plug-in:
  - Bug fix - the "Pre-Depends" field is added to the package dependencies.
  - Bug fix - not all packages have dependencies, so they inherited those of the
    previous package.
  - The update repositories were added.