Bug fix - when searching for packages, prefer perfect matches.
In some cases, package names appear both as the names of independent packages
and as alternate names of other packages. Most often, this happens with
backwards-compatibility wrappers.

This commit changes the behavior of roar-ng II: when package search is
performed, its results get prioritized according to the package entry accuracy.
If the package name given is an alternate name in a package entry, it is
considered less accurate, so entries with the given name as the package name
are always preferred.