Created a dedicated directory for configuration.
All distribution-related configuration (e.g package repositories and mirrors)
was moved to the new directory. This is required in order to make roar-ng II
more configurable.

In addition, this is a milestone towards the goal of moving package download
functionality to an external script, which will be part of distributions built
using roar-ng II.