Added package fetching support to hpm.
This support is based on roar-ng II's own code. This integration is possible
thanks to the new design of roar-ng II.

hpm was extended with two additional tools, hpm-makedb and hpm-fetch. The former
creates a package database and the latter downloads a package to the current
directory.

The list of distributions used by these tools is configurable, through
/etc/roar-ng/distro.list, while the package repositories and mirrors are
configurable through /etc/roar-ng/distro.