Bug fix - module dependency files are not deleted.
In addition, the symlinks to the kernel sources directory are placed in the
wrong directory.

These problems are caused by the kernel build script, which still uses the
deprecated PKG_MAJOR_VER, which is undefined. If the depmod implementation the
package was built against is incompatible with the one used by udev, this
prevents modules from being loaded.